EPBDS Updating after POI changes.

+++ POI/trunk/org.openl.lib.poi35.dev.modified/src/org/apache/poi/hssf/record/formula/functions/ArrayMode.java
@@ -11,6 +11,6 @@ import org.apache.poi.hssf.record.formula.eval.ValueEval;
  */
 public interface ArrayMode {
 
-	ValueEval evaluateInArrayFormula(ValueEval[] args, int srcRowIndex, short srcColumnIndex);
+	ValueEval evaluateInArrayFormula(ValueEval[] args, int srcRowIndex, int srcColumnIndex);
 
 }
